+++ api/migrations/2026-09-13-120000-0000_add_display_date_formats/down.sql
@@ -1,0 +1,3 @@
+ALTER TABLE app_settings
+DROP COLUMN datetime_format,
+DROP COLUMN date_format;