+++ api/migrations/2026-09-13-120000-0000_add_display_date_formats/up.sql
@@ -1,0 +1,3 @@
+ALTER TABLE app_settings
+ADD COLUMN date_format TEXT NOT NULL DEFAULT 'yyyy-MM-dd',
+ADD COLUMN datetime_format TEXT NOT NULL DEFAULT 'MM/dd/yyyy HH:mm';